reorganisation des pages
