--- client/src/components/Beers.ts
@@ -1,80 +1,0 @@
-// export interface Beer {
-//   id: number;
-//   name: string;
-//   brewery_type: string;
-//   address: string;
-//   city: string;
-//   state_province: string;
-//   postal_code: string;
-//   country: string;
-//   website_url: string;
-//   state: string;
-// }
-
-// export const beersList: Beer[] = [
-//   {
-//     id: 1,
-//     name: "Bute Brew Co",
-//     brewery_type: "closed",
-//     address: "15-17 Columshill Street",
-//     city: "Isle of Bute",
-//     state_province: "Bute",
-//     postal_code: "PA20 0DN",
-//     country: "Scotland",
-//     website_url: "http://butebrewco.co.uk/",
-//     state: "Bute",
-//   },
-//   {
-//     id: 2,
-//     name: "Colonsay Ales",
-//     brewery_type: "micro",
-//     address: "Colonsay Brewery",
-//     city: "Isle of Colonsay",
-//     state_province: "Argyll",
-//     postal_code: "PA61 7YZ",
-//     country: "Scotland",
-//     website_url: "https://colonsaybrewery.co.uk/",
-//     state: "Argyll",
-//   },
-//   {
-//     id: 3,
-//     name: "Fyne Ales",
-//     brewery_type: "micro",
-//     address: "Achadunan",
-
-//     city: "Cairndow",
-//     state_province: "Argyll",
-//     postal_code: "PA26 8BJ",
-//     country: "Scotland",
-
-//     website_url: "https://www.fyneales.com/",
-//     state: "Argyll",
-//   },
-//   {
-//     id: 4,
-//     name: "Glen Luss Craft Brewery & Distillery",
-//     brewery_type: "micro",
-//     address: "Church Road",
-
-//     city: "Alexandria",
-//     state_province: "West Dunbartonshire",
-//     postal_code: "G83 8NZ",
-//     country: "Scotland",
-
-//     website_url: "https://glenluss.co.uk/",
-//     state: "West Dunbartonshire",
-//   },
-//   {
-//     id: 5,
-//     name: "Islay Ales",
-//     brewery_type: "micro",
-//     address: "Islay House Square",
-
-//     city: "Isle of Islay",
-//     state_province: "Argyll",
-//     postal_code: "PA44 7NZ",
-//     country: "Scotland",
-//     website_url: "https://www.islayales.com/",
-//     state: "Argyll",
-//   },
-// ];
