--- client/src/components/BeersList.tsx
@@ -1,18 +1,0 @@
-// import BeerCard from "./BeerCard";
-// import type { Beer } from "./Beers";
-
-// interface BeersListProps {
-//   beers: Beer[];
-// }
-
-// function BeersList({ beers }: BeersListProps) {
-//   return (
-//     <div>
-//       {beers.map((beer: Beer) => (
-//         <BeerCard key={beer.id} beer={beer} />
-//       ))}
-//     </div>
-//   );
-// }
-
-// export default BeersList;
